refactor: Refactor app into module
